refactor: split align input box
