+++ src/components/editor/sidebar-element-settings-tab/layout-setting.tsx
@@ -1,17 +1,13 @@
 "use client";
 
 import {
-  AlignCenterIcon,
-  AlignLeftIcon,
-  AlignRight,
   ArrowDownIcon,
   ArrowRightIcon,
   CornerDownLeftIcon,
-  DotIcon,
-  type LucideProps,
 } from "lucide-react";
 import { useMemo, useState } from "react";
 import { useEditor } from "~/components/editor/provider";
+import AlignInputBox from "~/components/editor/ui/align-input-box";
 import { IconInput } from "~/components/editor/ui/input";
 import {
   GapIcon,
@@ -26,7 +22,6 @@ import {
 import { Toggle } from "~/components/ui/toggle";
 import { ToggleGroup, ToggleGroupItem } from "~/components/ui/toggle-group";
 import TooltipSimple from "~/components/ui/tooltip-simple";
-import { cn } from "~/lib/utils";
 
 export default function LayoutSetting() {
   return (
@@ -36,122 +31,6 @@ export default function LayoutSetting() {
       </div>
       <FlexBoxSection />
       <PaddingSection />
-    </div>
-  );
-}
-
-const alignConfig = {
-  start_start: {
-    style: {
-      justifyContent: "flex-start",
-      alignItems: "flex-start",
-    },
-    Icon: (props: LucideProps) => <AlignLeftIcon {...props} />,
-  },
-  start_center: {
-    style: {
-      justifyContent: "flex-start",
-      alignItems: "center",
-    },
-    Icon: (props: LucideProps) => <AlignCenterIcon {...props} />,
-  },
-  start_end: {
-    style: {
-      justifyContent: "flex-start",
-      alignItems: "flex-end",
-    },
-    Icon: (props: LucideProps) => <AlignRight {...props} />,
-  },
-  center_start: {
-    style: {
-      justifyContent: "center",
-      alignItems: "flex-start",
-    },
-    Icon: (props: LucideProps) => <AlignLeftIcon {...props} />,
-  },
-  center_center: {
-    style: {
-      justifyContent: "center",
-      alignItems: "center",
-    },
-    Icon: (props: LucideProps) => <AlignCenterIcon {...props} />,
-  },
-  center_end: {
-    style: {
-      justifyContent: "center",
-      alignItems: "flex-end",
-    },
-    Icon: (props: LucideProps) => <AlignRight {...props} />,
-  },
-  end_start: {
-    style: {
-      justifyContent: "flex-end",
-      alignItems: "flex-start",
-    },
-    Icon: (props: LucideProps) => <AlignLeftIcon {...props} />,
-  },
-  end_center: {
-    style: {
-      justifyContent: "flex-end",
-      alignItems: "center",
-    },
-    Icon: (props: LucideProps) => <AlignCenterIcon {...props} />,
-  },
-  end_end: {
-    style: {
-      justifyContent: "flex-end",
-      alignItems: "flex-end",
-    },
-    Icon: (props: LucideProps) => <AlignRight {...props} />,
-  },
-} as const;
-
-function AlignInput({
-  value,
-  onChangeValue,
-}: {
-  value: { justifyContent?: string; alignItems?: string };
-  onChangeValue: (value: {
-    justifyContent: string;
-    alignItems: string;
-  }) => void;
-}) {
-  return (
-    <div className="grid grid-cols-3 grid-rows-3 rounded-sm border text-muted-foreground">
-      {Object.keys(alignConfig).map((key) => {
-        const alignKey = key as keyof typeof alignConfig;
-        const { Icon } = alignConfig[alignKey];
-        const [alignItemValue, justifyValue] = alignKey.split("_");
-        const isSelected =
-          value.alignItems === alignItemValue &&
-          value.justifyContent === justifyValue;
-
-        return (
-          <div
-            key={alignKey}
-            className="group relative flex h-5 w-5 items-center justify-center"
-            onClick={() =>
-              onChangeValue({
-                alignItems: alignItemValue,
-                justifyContent: justifyValue,
-              })
-            }
-          >
-            <DotIcon
-              size={14}
-              className={cn("group-hover:hidden", isSelected && "hidden")}
-            />
-            <div
-              className={cn(
-                "absolute inset-0 hidden items-center justify-center group-hover:flex",
-                isSelected && "flex text-foreground",
-              )}
-            >
-              <Icon size={16} />
-            </div>
-          </div>
-        );
-      })}
     </div>
   );
 }
@@ -253,7 +132,7 @@ function FlexBoxSection() {
         />
       </div>
       <div className="col-span-4 row-span-2 flex items-start">
-        <AlignInput
+        <AlignInputBox
           value={{
             alignItems: element.styles.alignItems,
             justifyContent: element.styles.justifyContent,
